Revert "feat: Observation Details Responsiveness, WEB-983 (#4958)"
This reverts commit 46ab9e1cc806a19fae218631481cee28db107077.
