+++ app/assets/stylesheets/autocomplete.scss
@@ -95,25 +95,6 @@ $label-start: 7px;
     max-width: calc( 100vw - 90px );
   }
 }
-
-@media ( max-width: $md-max ) {
-  .ui-autocomplete {
-    max-width: 100vw;
-    .ac {
-      min-width: 0;
-    }
-    &.projects {
-      width: auto;
-      .ac {
-        width: auto;
-        min-width: 0;
-      }
-    }
-  }
-  .ui-autocomplete.taxon-autocomplete {
-    max-width: 100vw;
-  }
-}
 .ac-message[data-type='message'] {
   cursor: default;
 }
