+++ app/assets/stylesheets/index.scss
@@ -2180,7 +2180,7 @@ $header-line-height: 34px;
     opacity: 1;
   }
   .navtabs {
-    margin-right: clamp(5px, 3.5vw, 20px);
+    margin-right: 20px;
     padding: 10px 0px;
     line-height: 27px;
     .navtab {
