+++ app/assets/stylesheets/shared/user_text.scss
@@ -12,13 +12,9 @@
     position: relative;
     overflow: hidden;
     transition: height 2s;
-    overflow-wrap: anywhere;
+    overflow-wrap: break-word;
     img {
       max-width: 100%;
-    }
-    pre {
-      max-width: 100%;
-      overflow-x: auto;
     }
     > p:last-child,
     > ul:last-child,
