+++ app/assets/stylesheets/shared/user_with_icon.scss
@@ -9,12 +9,8 @@
     flex-grow: 2;
     display: flex;
     flex-direction: column;
-    min-width: 0; // allow the flex column to shrink below content width so .title can ellipsize
   }
   .title {
-    overflow: hidden;
-    text-overflow: ellipsis;
-    white-space: nowrap;
     a {
       color: #000;
     }
