+++ app/controllers/observations_controller.rb
@@ -328,8 +328,10 @@ class ObservationsController < ApplicationController
 
         @skip_application_js = true
         @flash_js = true
-        @responsive = true
-        @skip_min_width = true
+        if current_user&.in_test_group?( "responsive-header" ) &&
+            current_user.in_test_group?( "responsive-obs-detail" )
+          @skip_min_width = true
+        end
         return render layout: "bootstrap"
       end
