+++ app/webpack/observations/show/components/assessment.jsx
@@ -3,6 +3,7 @@
 import React from "react";
 import PropTypes from "prop-types";
 import {
+  Grid,
   Row,
   Col,
   Panel,
@@ -30,6 +31,8 @@ class Assessment extends React.Component {
     } = this.props;
     if ( !observation ) { return ( <span /> ); }
     const loggedIn = config && config.currentUser;
+    const responsive = config.currentUser?.isAdmin
+      && config.currentUser?.isInTestGroup( "responsive-obs-detail" );
     const { open } = this.state;
     const InnerWrapper = innerWrapper || ( props => (
       <div>
@@ -40,7 +43,7 @@ class Assessment extends React.Component {
       </div>
     ) );
     return (
-      <div className="assessment_container">
+      <Grid>
         <div className="QualityMetrics collapsible-section">
           <div>
             <h3 className="collapsible">
@@ -189,21 +192,19 @@ class Assessment extends React.Component {
           <Panel expanded={open} onToggle={() => {}}>
             <Panel.Collapse>
               <InnerWrapper config={config}>
-                <div className="assessment_metrics">
-                  <Row>
-                    <Col sm={7}>
-                      <QualityMetricsContainer />
-                    </Col>
-                    <Col sm={5}>
-                      <ResearchGradeProgressContainer />
-                    </Col>
-                  </Row>
-                </div>
+                <Row>
+                  <Col xs={responsive ? null : 7} sm={responsive ? 7 : null}>
+                    <QualityMetricsContainer />
+                  </Col>
+                  <Col xs={responsive ? null : 5} sm={responsive ? 5 : null}>
+                    <ResearchGradeProgressContainer />
+                  </Col>
+                </Row>
               </InnerWrapper>
             </Panel.Collapse>
           </Panel>
         </div>
-      </div>
+      </Grid>
     );
   }
 }
