+++ app/webpack/shared/components/carousel.tsx
@@ -3,8 +3,6 @@ import React, {
 } from "react";
 import css from "./carousel.module.css";
 
-const ITEM_GAP_SIZE = 5; // px
-
 export interface CarouselProps {
   items: React.ReactElement[];
   finalItem?: React.ReactElement;
@@ -28,12 +26,13 @@ const Carousel = ( {
   const [trackWidth, setTrackWidth] = useState( 0 );
   const [itemWidth, setItemWidth] = useState( 0 );
   const trackRef = useRef<HTMLDivElement>( null );
-  const allItems = finalItem ? [...items, finalItem] : items;
 
   const measureItemWidth = useCallback( () => {
     if ( !trackRef.current ) return;
-    const children = Array.from( trackRef.current.children ) as HTMLElement[];
-    const child = children.find( el => !el.style.width && el.getBoundingClientRect().width > 0 );
+    const child = Array.from( trackRef.current.children ).find( el => {
+      const h = el as HTMLElement;
+      return !h.style.width && h.getBoundingClientRect().width > 0;
+    } ) as HTMLElement | undefined;
     if ( child ) {
       const w = child.getBoundingClientRect().width;
       setItemWidth( prev => ( prev !== w ? w : prev ) );
@@ -70,16 +69,19 @@ const Carousel = ( {
 
   useEffect( () => {
     const track = trackRef.current;
-    const stride = itemWidth ? itemWidth + ITEM_GAP_SIZE : 0;
     let raf: number;
     const handleScroll = () => {
       cancelAnimationFrame( raf );
       raf = requestAnimationFrame( () => {
-        if ( !track || !stride ) return;
-        // items are uniform width + gap, so the item nearest the left edge is
-        // just scrollLeft / stride — one cheap read instead of measuring each child.
-        const idx = Math.round( track.scrollLeft / stride );
-        setActiveIndex( Math.min( Math.max( idx, 0 ), allItems.length - 1 ) );
+        if ( !track ) return;
+        const trackLeft = track.getBoundingClientRect().left;
+        let closest = 0;
+        let minDist = Infinity;
+        Array.from( track.children ).forEach( ( child, i ) => {
+          const dist = Math.abs( child.getBoundingClientRect().left - trackLeft );
+          if ( dist < minDist ) { minDist = dist; closest = i; }
+        } );
+        setActiveIndex( closest );
       } );
     };
     if ( track ) track.addEventListener( "scroll", handleScroll, { passive: true } );
@@ -87,16 +89,16 @@ const Carousel = ( {
       if ( track ) track.removeEventListener( "scroll", handleScroll );
       cancelAnimationFrame( raf );
     };
-  }, [itemWidth, allItems.length] );
+  }, [] );
 
   // Until the carousel has been measured, render only a couple items so off-screen
   // elements don't mount and fire network requests on first paint. After ResizeObserver
   // measures the track, visibleCount switches to the real value and placeholders get sized.
   const INITIAL_VISIBLE = 2;
   const visibleCount = trackWidth && itemWidth
-    ? Math.ceil( trackWidth / ( itemWidth + ITEM_GAP_SIZE ) )
+    ? Math.ceil( trackWidth / itemWidth )
     : INITIAL_VISIBLE;
-  const allItemsVisible = allItems.length < visibleCount;
+  const allItems = finalItem ? [...items, finalItem] : items;
   const renderedItems = allItems.map( ( item, index ) => {
     const distFromActive = Math.abs( index - activeIndex );
     return (
@@ -131,7 +133,7 @@ const Carousel = ( {
         <p className="text-muted text-center">{ noContent }</p>
       ) }
       <div className={css.body}>
-        { allItems.length > 0 && !allItemsVisible && (
+        { allItems.length > 0 && (
           <button
             type="button"
             className={`btn ${css.navbtn}`}
@@ -142,10 +144,10 @@ const Carousel = ( {
             ❮
           </button>
         ) }
-        <div className={allItemsVisible ? css.statictrack : css.track} ref={trackRef}>
+        <div className={css.track} ref={trackRef}>
           { renderedItems }
         </div>
-        { allItems.length > 0 && !allItemsVisible && (
+        { allItems.length > 0 && (
           <button
             type="button"
             className={`btn ${css.navbtn}`}
