+++ app/webpack/shared/components/rtl_test_group_toggle.jsx
@@ -12,12 +12,18 @@ const RtlTestGroupToggle = ( { config } ) => {
     return null;
   }
   return (
-    <TestGroupToggle
-      group="rtl"
-      joinPrompt={I18n.t( "rtl_test_prompt" )}
-      joinedStatus={I18n.t( "rtl_test_joined_status" )}
-      user={config.currentUser}
-    />
+    <div className="container">
+      <div className="row">
+        <div className="cols-xs-12">
+          <TestGroupToggle
+            group="rtl"
+            joinPrompt={I18n.t( "rtl_test_prompt" )}
+            joinedStatus={I18n.t( "rtl_test_joined_status" )}
+            user={config.currentUser}
+          />
+        </div>
+      </div>
+    </div>
   );
 };
