+++ app/webpack/shared/components/taxon_thumbnail.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import CoverImage from "./cover_image";
 import SplitTaxon from "./split_taxon";
 import css from "./taxon_thumbnail.module.css";
-import type { Taxon, Config, Photo } from "../types";
+import type { Taxon, Config } from "../types";
 import { urlForTaxon as baseUrlForTaxon } from "../../taxa/shared/util";
 
 // baseUrlForTaxon returns `string | null`; coerce to the prop contract
@@ -18,9 +18,6 @@ export interface TaxonThumbnailBadge {
 
 export interface TaxonThumbnailProps {
   taxon: Taxon;
-  // undefined falls back to the taxon's default photo; null suppresses that
-  // fallback and forces the iconic placeholder (e.g. for a photoless observation).
-  photo?: Photo | null;
   width?: number;
   className?: string;
   overlay?: React.ReactNode;
@@ -34,7 +31,6 @@ export interface TaxonThumbnailProps {
 
 const TaxonThumbnail = ( {
   taxon,
-  photo,
   width,
   className,
   overlay,
@@ -47,19 +43,12 @@ const TaxonThumbnail = ( {
 }: TaxonThumbnailProps ) => {
   let mediumURL: string | undefined;
   let squareURL: string | undefined;
-  if ( photo && typeof photo.photoUrl === "function" ) {
-    mediumURL = photo.photoUrl( "medium" );
-    squareURL = photo.photoUrl( "square" );
-  } else if ( photo !== null ) {
-    // only fall back to the taxon's default photo when no photo was passed;
-    // an explicit null means "this thing has no photo, show the placeholder".
-    if ( taxon.defaultPhoto && typeof taxon.defaultPhoto.photoUrl === "function" ) {
-      mediumURL = taxon.defaultPhoto.photoUrl( "medium" );
-      squareURL = taxon.defaultPhoto.photoUrl( "square" );
-    } else if ( taxon.default_photo && typeof taxon.default_photo === "object" ) {
-      mediumURL = taxon.default_photo.medium_url;
-      squareURL = taxon.default_photo.square_url;
-    }
+  if ( taxon.defaultPhoto && typeof taxon.defaultPhoto.photoUrl === "function" ) {
+    mediumURL = taxon.defaultPhoto.photoUrl( "medium" );
+    squareURL = taxon.defaultPhoto.photoUrl( "square" );
+  } else if ( taxon.default_photo && typeof taxon.default_photo === "object" ) {
+    mediumURL = taxon.default_photo.medium_url;
+    squareURL = taxon.default_photo.square_url;
   }
 
   const wrapperStyle: React.CSSProperties | undefined = width
