--- spec/e2e/app_commands/make_curator.rb
@@ -1,5 +1,0 @@
-opts = command_options.is_a?( Hash ) ? command_options.transform_keys( &:to_sym ) : {}
-user = User.make!( opts )
-role = Role.find_by( name: "curator" ) || Role.create!( name: "curator" )
-user.roles << role
-user
