+++ spec/e2e/playwright/fixtures/observation-response.ts
@@ -46,8 +46,7 @@ export function buildObservationApiResponse( obsData: Record<string, unknown> ) 
         user: {
           id: obsData["user_id"],
           login: "e2e_test_user",
-          name: "E2E Test User",
-          preferences: {}
+          name: "E2E Test User"
         },
         photos: [
           {
@@ -64,9 +63,7 @@ export function buildObservationApiResponse( obsData: Record<string, unknown> ) 
         spam: false,
         captive: false,
         out_of_range: false,
-        comments: ( obsData["comments"] as unknown[] ) || [],
-        ofvs: ( obsData["ofvs"] as unknown[] ) || [],
-        comments_count: ( ( obsData["comments"] as unknown[] ) || [] ).length,
+        comments_count: 0,
         identifications_count: 1,
         faves_count: 0,
         num_identification_agreements: 1,
