+++ spec/e2e/playwright/helpers/overflow.helper.ts
@@ -11,8 +11,8 @@ async function measureHorizontalOverflow(
 }
 
 export function expectNoHorizontalOverflow(
-  path: string | ( () => string ) = "/",
-  options: { waitForSelector?: string; setup?: ( page: Page ) => Promise<void> } = {}
+  path = "/",
+  options: { waitForSelector?: string } = {}
 ): void {
   test.describe( "no horizontal overflow", () => {
     for ( const name of Object.keys( VIEWPORTS ) as BreakpointName[] ) {
@@ -20,9 +20,8 @@ export function expectNoHorizontalOverflow(
 
       test( `the document body does not overflow the viewport at the ${name} breakpoint (${viewport.width}px)`, async ( { page } ) => {
         await page.setViewportSize( viewport );
-        if ( options.setup ) { await options.setup( page ); }
-        await page.goto( typeof path === "function" ? path() : path );
-        await page.locator( options.waitForSelector || "#header" ).first().waitFor();
+        await page.goto( path );
+        await page.locator( options.waitForSelector || "#header" ).waitFor();
 
         const { bodyWidth, viewportWidth } = await measureHorizontalOverflow( page );
